pyproject.toml
[tool.poetry]
name = "contacts-api"
version = "0.1.0"
description = "A FastAPI-based contacts management API"
authors = ["Your Name <youremail@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = "^2.0.15"
asyncpg = "^0.27.0"
pydantic-settings = "^2.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
